+++ CliMenu/Components/ButtonBehaviors/ButtonDisplayer.cs
@@ -1,0 +1,8 @@
+using CliMenu.Components.Behaviors;
+
+namespace CliMenu.Components.ButtonBehaviors;
+
+public class ButtonDisplayer : Displayer<Button>
+{
+    protected override string ProtectedGetDisplay(Button component) => $"[{component.Text}]";
+}