+++ CliMenu/Interfaces/IFocusManager.cs
@@ -1,22 +1,22 @@
 namespace CliMenu.Interfaces;
 
 /// <summary>
-/// Interface defining Focus Manager behaviour.
+/// Manages the focus state of a component.
 /// </summary>
 public interface IFocusManager
 {
     /// <summary>
-    /// Gets or sets a value indicating whether the component is currently focused.
+    /// Gets a value indicating whether the component is currently focused.
     /// </summary>
     bool IsFocused { get; }
 
     /// <summary>
-    /// Sets <see cref="IsFocused"/> to true and has aditional logic.
+    /// Sets focus to the component.
     /// </summary>
     void Focus();
 
     /// <summary>
-    /// Sets <see cref="IsFocused"/> to false and has aditional logic.
+    /// Removes focus from the component.
     /// </summary>
     void Blur();
 }