+++ CliMenu/Components/DefaultComponents/BarBehaviors/BarDisplayer.cs
@@ -1,0 +1,11 @@
+using CliMenu.Components.Behaviors;
+
+namespace CliMenu.Components.DefaultComponents.BarBehaviors;
+
+public class BarDisplayer : Displayer<Bar>
+{
+    protected override string ProtectedGetDisplay(Bar component)
+    {
+        throw new NotImplementedException();
+    }
+}